# Avoid expensive operations in CaptureEvent hot path