--- a/client_test.go
+++ b/client_test.go
@@ -2,6 +2,9 @@ package sentry
 
 import (
 	"errors"
+	"fmt"
+	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -321,4 +324,60 @@ func TestBeforeSendGetAccessToEventHint(t *testing.T) {
 	client.CaptureException(ex, &EventHint{OriginalException: ex}, scope)
 
 	assertEqual(t, transport.lastEvent.Message, "customComplexError: Foo 42")
+}
+
+func TestSampleRate(t *testing.T) {
+	tests := []struct {
+		SampleRate float64
+		// tolerated range is [SampleRate-MaxDelta, SampleRate+MaxDelta]
+		MaxDelta float64
+	}{
+		// {0.00, 0.0}, // oddly, sample rate = 0.0 means 1.0, skip test for now.
+		{0.25, 0.2},
+		{0.50, 0.2},
+		{0.75, 0.2},
+		{1.00, 0.0},
+	}
+	for _, tt := range tests {
+		tt := tt
+		t.Run(fmt.Sprint(tt.SampleRate), func(t *testing.T) {
+			var (
+				captureMessageCount uint64
+				beforeSendCount     uint64
+			)
+			c, err := NewClient(ClientOptions{
+				SampleRate: tt.SampleRate,
+				BeforeSend: func(event *Event, hint *EventHint) *Event {
+					atomic.AddUint64(&beforeSendCount, 1)
+					return event
+				},
+			})
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			// Simulate using client from multiple hubs/goroutines to cover data
+			// races.
+			var wg sync.WaitGroup
+			mainHub := NewHub(c, NewScope())
+			for i := 0; i < 4; i++ {
+				wg.Add(1)
+				go func() {
+					defer wg.Done()
+					// hub shares the client with mainHub.
+					hub := mainHub.Clone()
+					for j := 0; j < 10000; j++ {
+						atomic.AddUint64(&captureMessageCount, 1)
+						hub.CaptureMessage("test")
+					}
+				}()
+			}
+			wg.Wait()
+
+			eventRate := float64(beforeSendCount) / float64(captureMessageCount)
+			if eventRate < tt.SampleRate-tt.MaxDelta || eventRate > tt.SampleRate+tt.MaxDelta {
+				t.Errorf("effective sample rate was %f, want %f±%f", eventRate, tt.SampleRate, tt.MaxDelta)
+			}
+		})
+	}
 }